refactor: Organize components into subdirectories and move utility functions to their respective features
