+++ src/app/(root)/page.tsx
@@ -1,7 +1,7 @@
 import PageArea from '@/components/layout/page-area';
 import Section from '@/components/layout/section';
-import Command from '@/features/spy/command';
-import Modes from '@/features/spy/modes';
+import Command from '@/features/spy/components/command';
+import Modes from '@/features/spy/components/modes';
 
 export default function Page() {
   return (
