+++ src/features/spy/components/command.tsx
